+++ src/pyscipopt/__init__.py
@@ -1,5 +1,10 @@
 __version__ = '2.2.3'
 
+# required for Python 3.8 on Windows
+import os
+if hasattr(os, 'add_dll_directory'):
+    os.add_dll_directory(os.path.join(os.getenv('SCIPOPTDIR'), 'bin'))
+
 # export user-relevant objects:
 from pyscipopt.Multidict import multidict
 from pyscipopt.scip      import Model