+++ src/pyscipopt/scip.pyx
@@ -3006,7 +3006,7 @@ cdef class Model:
             nsubproblems = SCIPbendersGetNSubproblems(_benders[i])
             for j in range(nsubproblems):
                 PY_SCIP_CALL(SCIPsetupBendersSubproblem(self._scip,
-                    _benders[i], self._bestSol.sol, j))
+                    _benders[i], self._bestSol.sol, j, SCIP_BENDERSENFOTYPE_CHECK))
                 PY_SCIP_CALL(SCIPsolveBendersSubproblem(self._scip,
                     _benders[i], self._bestSol.sol, j, &_infeasible, solvecip, NULL))
 
@@ -3075,11 +3075,13 @@ cdef class Model:
         """
         SCIPbendersSetSubproblemIsConvex(benders._benders, probnumber, isconvex)
 
-    def setupBendersSubproblem(self, probnumber, Benders benders = None, Solution solution = None):
+    def setupBendersSubproblem(self, probnumber, checktype, Benders benders = None, Solution solution = None):
         """ sets up the Benders' subproblem given the master problem solution
 
         Keyword arguments:
         probnumber -- the index of the problem that is to be set up
+        checktype -- the type of solution check that prompted the solving of the Benders' subproblems, either
+            PY_SCIP_BENDERSENFOTYPE: LP, RELAX, PSEUDO or CHECK
         benders -- the Benders' decomposition to which the subproblem belongs to
         solution -- the master problem solution that is used for the set up, if None, then the LP solution is used
         """
@@ -3096,7 +3098,7 @@ cdef class Model:
         else:
             scip_benders = benders._benders
 
-        retcode = SCIPsetupBendersSubproblem(self._scip, scip_benders, scip_sol, probnumber)
+        retcode = SCIPsetupBendersSubproblem(self._scip, scip_benders, scip_sol, probnumber, checktype)
 
         PY_SCIP_CALL(retcode)