+++ src/pyscipopt/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '3.0.0'
+__version__ = '3.0.1'
 
 # required for Python 3.8 on Windows
 import os