+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - define Python object identity based on underlying SCIP object pointers, so that e.g. rows and columns can be added to sets, and testing for equality is consistent over time.
 - add Row.isRemovable and Row.getOrigintype
 - remove Python 2.7 support
+- fix documentation errors
 
 ## 2.2.3 - 2019-12-10
 ### Added