+++ CHANGELOG.md
@@ -1,10 +1,18 @@
 # CHANGELOG
 
 ## Unreleased
+
+## 3.0.1 - 2020-07-05
+### Added
+- expose even more SCIP functionality in `scip.pxd`
+
 ### Changed
 - `Model.from_ptr` and `Model.to_ptr` use a `PyCapsule` to exchange the SCIP pointer
   rather than an integer.
 - mark getDualMultiplier() as deprecated, only getDualSolLinear() is supposed to be used to get duals of constraints
+
+### Removed
+* removed `__div__` from Expr and GenExpr to make it compatible with cython 0.29.20
 
 ## 3.0.0 - 2020-04-11
 ### Added